Refactor admin dashboard components: renamed headers in AdminDataTable for consistency, replaced ReviewDialog with IndexDialog, and removed obsolete table.tsx file. Updated AdminPanel to utilize the new _table component structure.
